add auth site
typesafe-i18n is not used currently, but installed.
Some config files were deleted, the database types were added, SignIn and SignUp components as well as the utility components `Card`, `CenterX` and `CenterY` and the component `Protect`, that uses useAuth() to redirect from protected sites if the user is not logged in, were added. The Auth page, that handles singin and signup was added. A card layout was created and the first routes are set up